requirements.txt
annotated-types==0.7.0
attrs==25.3.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
colormath==3.0.0
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
filetype==1.2.0
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.57.0
fsspec==2025.3.2
google-ai-generativelanguage==0.6.15
google-api-core==2.25.0rc0
google-api-python-client==2.168.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.5
googleapis-common-protos==1.70.0
grpcio==1.71.0
grpcio-status==1.71.0
httplib2==0.22.0
humanfriendly==10.0
idna==3.7
imageio==2.37.0
itsdangerous==2.2.0
Jinja2==3.1.6
jsonschema==4.23.0
jsonschema-specifications==2025.4.1
kiwisolver==1.4.8
lazy_loader==0.4
llvmlite==0.44.0
MarkupSafe==3.0.2
matplotlib==3.10.1
mpmath==1.3.0
networkx==3.4.2
numba==0.61.2
numpy==2.1.1
onnxruntime==1.21.1
opencv-python==4.11.0.86
opencv-python-headless==4.10.0.84
packaging==25.0
pandas==2.2.3
pillow==11.2.1
pillow_heif==0.22.0
platformdirs==4.3.7
pooch==1.8.2
proto-plus==1.26.1
protobuf==5.29.4
psutil==7.0.0
py-cpuinfo==9.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.3
pydantic_core==2.33.1
PyMatting==1.1.13
pyparsing==3.2.3
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
rembg==2.0.65
requests==2.32.3
requests-toolbelt==1.0.0
roboflow==1.1.61
rpds-py==0.24.0
rsa==4.9.1
scikit-image==0.25.2
scipy==1.15.2
seaborn==0.13.2
setuptools==79.0.1
six==1.17.0
sympy==1.13.3
tifffile==2025.3.30
torch==2.7.0
torchvision==0.22.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
ultralytics==8.3.115
ultralytics-thop==2.0.14
uritemplate==4.1.1
urllib3==2.4.0
Werkzeug==3.1.3
